Move addCommitToTriggerBuild() into ConversionCommandsTrait
